refactor(redux): refactor to be able to use multiple reducers with combineReducer
