+++ src/actions/templatesActions.js
