+++ src/reducers/templatesReducer.js
